refactor: 복잡한 추상화 없애고 재사용안되는 함수로 분리
