+++ src/app/mount/mountLoadMoreButton.js
@@ -2,5 +2,5 @@ import { $ } from "../../util/querySelector";
 
 export default function mountLoadMoreButton(loadMoreButton) {
   const $container = $("#thumbnail-container");
-  $container?.append(loadMoreButton.$el);
+  $container?.append(loadMoreButton);
 }
