+++ src/components/longButton/longButton.js
@@ -3,21 +3,13 @@ import { defaultOptions, defaultQueryObject } from "../../setting/settings";
 import { hideElement } from "../../view/InputView";
 import Button from "../button/button";
 
-export default function LongButton(text, load) {
+export default function LoadMoreButton({ text, onClick }) {
   const $longButton = Button({
     className: ["primary", "width-100"],
     placeholder: text,
     id: "load-more",
-    onClick: load,
+    onClick,
   });
 
-  function setOnClick(onClick) {
-    $longButton.onclick = onClick;
-  }
-
-  function hide() {
-    hideElement($longButton);
-  }
-
-  return { $el: $longButton, setOnClick, hide };
+  return $longButton;
 }
